app/requirements.txt
# Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
python-email-validator==2.1.0
dnspython==2.4.2
pytest==7.4.3
pytest-asyncio==0.21.1
aiofiles==23.2.1
python-magic==0.4.27
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
httpx==0.25.2

# Development dependencies (optional)
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0